PostGame Erika Bug Fix
Fixes the scenario where the player saves next to the stumps after beating Erika. Without redrawingmapview, the stumps would be trees until you walked away and walked back.

Redrawmapview is avoided if wXCoord == $25 so there is not a lag when running between route28 and mtsilver. It is not necessary as the tiles are off screen.