Hardmode: Pokecenter charge balance
After playing this for a bit, I realize the amount a pokecenter charges for a heal needs to be less than the cost of a pokeball, or the player is incentivized to zero out their cash stack with pokeballs to get a free/discount heal.
I mean, they still are, but that will be a bit trickier to do if the cost of the heal is less than the cheapest always useful item.